feat: gunzip reader to replace golang native
